refactor: Promote (Gear)TooltipModel into an upper-level package
