+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -28,7 +28,6 @@ import com.wynntils.models.dungeon.DungeonModel;
 import com.wynntils.models.elements.ElementModel;
 import com.wynntils.models.emeralds.EmeraldModel;
 import com.wynntils.models.gear.GearModel;
-import com.wynntils.models.gear.GearTooltipModel;
 import com.wynntils.models.horse.HorseModel;
 import com.wynntils.models.ingredients.IngredientModel;
 import com.wynntils.models.items.ItemModel;
@@ -51,6 +50,7 @@ import com.wynntils.models.territories.GuildAttackTimerModel;
 import com.wynntils.models.territories.TerritoryModel;
 import com.wynntils.models.territories.TowerAuraTimerModel;
 import com.wynntils.models.token.TokenModel;
+import com.wynntils.models.tooltip.GearTooltipModel;
 import com.wynntils.models.trademarket.TradeMarketModel;
 import com.wynntils.models.war.WarModel;
 import com.wynntils.models.worlds.BombModel;
