+++ common/src/main/java/com/wynntils/features/tooltips/ItemStatInfoFeature.java
@@ -12,9 +12,6 @@ import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.Config;
 import com.wynntils.core.persisted.config.ConfigCategory;
 import com.wynntils.mc.event.ItemTooltipRenderEvent;
-import com.wynntils.models.gear.tooltip.GearTooltipBuilder;
-import com.wynntils.models.gear.tooltip.GearTooltipStyle;
-import com.wynntils.models.gear.tooltip.TooltipIdentificationDecorator;
 import com.wynntils.models.gear.type.GearInfo;
 import com.wynntils.models.gear.type.GearInstance;
 import com.wynntils.models.items.WynnItemData;
@@ -23,6 +20,9 @@ import com.wynntils.models.stats.StatCalculator;
 import com.wynntils.models.stats.type.StatActualValue;
 import com.wynntils.models.stats.type.StatListOrdering;
 import com.wynntils.models.stats.type.StatPossibleValues;
+import com.wynntils.models.tooltip.GearTooltipBuilder;
+import com.wynntils.models.tooltip.GearTooltipStyle;
+import com.wynntils.models.tooltip.TooltipIdentificationDecorator;
 import com.wynntils.utils.mc.ComponentUtils;
 import com.wynntils.utils.mc.KeyboardUtils;
 import com.wynntils.utils.mc.McUtils;
