+++ common/src/main/java/com/wynntils/models/items/FakeItemStack.java
@@ -7,8 +7,8 @@ package com.wynntils.models.items;
 import com.wynntils.core.components.Handlers;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
-import com.wynntils.models.gear.tooltip.GearTooltipBuilder;
 import com.wynntils.models.items.items.game.GearItem;
+import com.wynntils.models.tooltip.GearTooltipBuilder;
 import java.util.List;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
