+++ common/src/main/java/com/wynntils/models/tooltip/GearTooltipHeader.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.gear.tooltip;
+package com.wynntils.models.tooltip;
 
 import com.wynntils.core.components.Models;
 import com.wynntils.models.activities.quests.QuestInfo;
